+++ crates/rome_js_parser/test_data/inline/err/export_default_expression_broken.js
@@ -1,0 +1,4 @@
+export default ;
+export default @decorator
+export default
+export default @decorator